requirements.txt
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3
django==2.2.16
django-filter==21.1
djangorestframework==3.11.2
djangorestframework-simplejwt==5.2.0
channels==4.0.0
Pillow==10.0.0